+++ Gw2Sharp/WebApi/V2/Models/Stories/StoryChapter.cs
@@ -1,0 +1,13 @@
+namespace Gw2Sharp.WebApi.V2.Models
+{
+    /// <summary>
+    /// Represents a story chapter.
+    /// </summary>
+    public class StoryChapter
+    {
+        /// <summary>
+        /// The story chapter name.
+        /// </summary>
+        public string Name { get; set; } = string.Empty;
+    }
+}